Add files via upload
Python utility script written to solve real time project problem to save the effort and time for finding the existence of terminated employees in licenses assigned repository for recovery of licenses of terminated employees or role change within organization and licenses are left used. This utility has been used production and helped quick unassigning licenses and make available the license key to the new employee requiring the license. This has saved cost to organization for avoiding the purchase of new licenses by paying addition cost to vendor support.